+++ allink_core/djangocms_vid/cms_plugins.py
@@ -3,54 +3,38 @@ from django.utils.translation import ugettext_lazy as _
 from cms.plugin_base import CMSPluginBase
 
 from cms.plugin_pool import plugin_pool
-from .models import AllinkImagePlugin
-from .forms import AllinkImagePluginForm
+from .models import AllinkVidPlugin
+from .forms import AllinkVidPluginForm
 
 
 @plugin_pool.register_plugin
-class CMSAllinkImagePlugin(CMSPluginBase):
-    model = AllinkImagePlugin
-    name = _('Image')
+class CMSAllinkVidPlugin(CMSPluginBase):
+    model = AllinkVidPlugin
+    name = _('Video')
     module = _("allink")
-    form = AllinkImagePluginForm
+    form = AllinkVidPluginForm
 
     def get_fieldsets(self, request, obj=None):
         fieldsets = [
             (None, {
                 'fields': (
-                    'picture',
-                    'external_picture',
+                    'label',
+                    'poster',
+                    'embed_link',
                     'ratio',
+                    'template',
                 )
             }),
             (_('Advanced settings'), {
                 'classes': ('collapse',),
                 'fields': (
-                    'template',
-                    'caption_text',
                     'attributes',
                 )
             }),
-            (_('Link settings'), {
-                'classes': ('collapse',),
-                'fields': (
-                    ('link_url', 'link_page',),
-                    ('link_mailto', 'link_phone'),
-                    ('link_anchor', 'link_special'),
-                    'link_file',
-                    'link_target',
-                )
-            }),
-            (_('Wireframe settings'), {
-                'classes': ('collapse',),
-                'fields': (
-                    ('width', 'use_no_cropping'),
-                )
-            })
         ]
 
         return fieldsets
 
     def get_render_template(self, context, instance, placeholder):
-        template = 'djangocms_image/{}/content.html'.format(instance.template)
+        template = 'djangocms_vid/{}/content.html'.format(instance.template)
         return template